Rename README.md to Creating-new-Repositoty
Here I have added phyton codes for the following exercise. 
"Define two variables and assign them values of 100 and 29. After this, make the interpreter sum up the two numbers and multiply their result by 3.
Calculate the 2nd exponent of the previous number and save it to a new variable.
Make the program print the result in the following way:
The result of the calculation was: